requirements.txt
aiofiles==23.2.1
aiohttp==3.8.6
aiosignal==1.3.1
altair==5.1.2
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.7.2
astroid==2.15.5
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
blinker==1.7.0
botocore==1.31.73
Brotli==1.1.0
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
dataclasses-json==0.6.1
decorator==5.1.1
dill==0.3.7
Django==4.2.3
duckduckgo-search==3.9.4
exceptiongroup==1.1.3
executing==2.0.1
frozenlist==1.4.0
git-remote-codecommit==1.17
gitdb==4.0.11
GitPython==3.1.40
greenlet==3.0.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.1
httpx==0.25.1
hyperframe==6.0.1
idna==3.4
ikp3db==1.4.1
importlib-metadata==6.8.0
importlib-resources==6.1.0
ipython==8.12.3
isort==5.12.0
jedi==0.19.0
Jinja2==3.1.2
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
langchain==0.0.330
langsmith==0.0.57
lazy-object-proxy==1.9.0
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.24.4
openai==0.28.1
packaging==23.2
pandas==2.0.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.1.0
pkgutil-resolve-name==1.3.10
platformdirs==3.11.0
prompt-toolkit==3.0.39
protobuf==4.25.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==14.0.0
pydantic==2.4.2
pydantic-core==2.10.1
pydeck==0.8.1b0
Pygments==2.16.1
pylint==2.17.4
pylint-django==2.5.3
pylint-flask==0.6
pylint-plugin-utils==0.8.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
rich==13.6.0
rpds-py==0.12.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
socksio==1.0.0
soupsieve==2.5
SQLAlchemy==2.0.23
sqlparse==0.4.4
stack-data==0.6.3
streamlit==1.28.1
tenacity==8.2.3
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.1
toolz==0.12.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.13.0
typing-extensions==4.8.0
typing-inspect==0.9.0
tzdata==2023.3
tzlocal==5.2
urllib3==1.26.18
validators==0.22.0
watchdog==3.0.0
wcwidth==0.2.8
wikipedia==1.4.0
wrapt==1.15.0
yarl==1.9.2
zipp==3.17.0
